Extract assign viking field from raid form to AssignViking widget
