+++ lib/widgets/widgets.dart
@@ -8,3 +8,5 @@ export './raid_table.dart';
 export './raid_console.dart';
 export './raid_chat.dart';
 export './message.dart';
+export './date_selector.dart';
+export './assign_viking.dart';
